# Optimize QuadraticTrinomial methods by 4-100x