These timings are taken across 10000 runs for a randomly generated polynomial.

```
old discriminant: 0.3621s
new discriminant: 0.0026s
old complex_roots: 0.8682s
new complex_roots: 0.0061s
old real_factors: 1.4473s
new real_factors: 0.2728s
old complex_factors: 1.0949s
new complex_factors: 0.2666s
```

Accessing polynomial elements through self.a, self.b, etc is a lot slower than accessing self._vector directly - doing this allows substantial speedups without adding a lot of extra code.